pyproject.toml
[tool.poetry]
name = "home-sensors-data-server"
version = "0.0.1"
description = ""
authors = ["Gosha Simakov <yolo73104@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
loguru = "^0.7.2"
nameko = "^2.14.1"
fastapi-pagination = "^0.12.13"
prometheus-fastapi-instrumentator = "^6.1.0"
fastapi-async-sqlalchemy = "^0.6.0"
uvicorn = "^0.24.0.post1"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
aio-pika = "^9.3.1"
pika = "^1.3.2"
aiohttp = "^3.9.1"
httpx = "^0.25.2"
sqlmodel = "^0.0.14"
sqlalchemy-utils = "^0.41.1"
cryptography = "^42.0.0"
alembic = "^1.13.1"
psycopg = "^3.1.17"
psycopg-binary = "^3.1.17"
arprequest = "^0.3"
scapy = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"